Extracted logic to upload and download attachments to a helper class.
Issue: eclipse#426.

Attachments are supported by multiple entity types. Therefore, it
makes sense to have this functionality available at a central place.
